test: add testinfra tests for verify steps
Add testinfra tests for the verify step and also for a new verify step,
which tests the host after running the side_effect step for
projects_rev01.
The testinfra tests test that the directory structure and files are as
expected, that the containers running are as expected and that endpoints
return their expected values.