other: set valid namespace and role_name to make Molecule work
Molecule apparently computes the fully qualified role name, if it isn't
given explictly, and then checks if it is valid.
Since Molecule isn't happy with the computed one, explictly provide a
valid namespace and role_name, so that the resulting fully qualified
role name is valid and Molecule is happy.

The namespace and role name should work for usage with Ansible Galaxy,
however we're not using Ansible Galaxy.

Also trying to skip this role name check by adding role-name to the
skip_list in an ansible-lint config file, resulted in ansible-lint
indeed skipping the check, however Molecule would still check the role
name and fail (so even tho the resulting error message suggests that
using skip_list works, it apparently doesn't).